backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Face recognition
face_recognition==1.3.0
numpy==1.26.3

# HTTP client (for MSG91)
httpx==0.26.0

# Configuration
python-dotenv==1.0.0

# Validation
pydantic[email]==2.5.3

# Image processing (for admin tool)
Pillow==10.2.0

